refactor:remove unnecessary hierarchy directories
