+++ src/main/java/com/asclepius/controller/CardController.java
@@ -1,4 +1,4 @@
-package com.asclepius.controller.card;
+package com.asclepius.controller;
 
 import com.asclepius.common.ResponseCode;
 import com.asclepius.dto.CardDTO;
