+++ src/main/java/com/asclepius/controller/DoctorController.java
@@ -1,4 +1,4 @@
-package com.asclepius.controller.doctor;
+package com.asclepius.controller;
 
 import com.asclepius.dto.ResultDTO;
 import com.asclepius.service.DoctorService;
